Updating my setup further

+++ homebrew.sh
@@ -1,6 +1,8 @@
 #!/usr/bin/env bash
-
 # Install command-line tools using Homebrew.
+
+echo "Install Homebrew"
+/usr/bin/ruby -e "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install)"
 
 # Make sure we’re using the latest Homebrew.
 brew update
@@ -9,10 +11,18 @@ brew update
 brew upgrade
 
 # Install iTerm2
-brew cask install iterm2
+echo -n "Need iTerm? (y/n) "
+read REPLY
+if [[ $REPLY =~ ^[Yy]$ ]]; then
+    brew cask install iterm2
+fi;
 
 #Install Hyper.js
-brew cask install hyper
+echo -n "Need some hyper terminal action? (y/n) "
+read REPLY
+if [[ $REPLY =~ ^[Yy]$ ]]; then
+    brew cask install hyper
+fi;
 
 # Remove outdated versions from the cellar.
 brew cleanup
